\name{alarm.net}  
\alias{alarm.net}
\docType{data}
\title{ Graph structure of ALARM network}
\description{
  The adjacency matrix of the ALARM network graph structure.
}
\usage{data(alarm.net)}
\source{
  Beinlich, I., Suermondt, H., Chavez, R. and Cooper, G. (1989). The
  ALARM monitoring system: A case study with two probabilistic inference
  techniques for belief networks, in: \emph{Proceedings of the 2nd
    European Conference on Artificial Intelligence in Medicine}, pp.247-256,
  Springer-Verlag, Berlin.
}
\keyword{datasets} 
 