\name{all.equal-methods}  
\docType{methods}
\alias{all.equal-methods}
\alias{all.equal,sep.pair,sep.pair-method}
\title{All.equal method for sep.pair class} 
\description{
 Checks if two \code{sep.pair} objects are all.equal. We call two
 \code{sep.pair} objects all.equal if they have the same (\code{u},\code{v})
 set, regardless of order.
}
\section{Methods}{
\describe{

\item{target = "sep.pair", current = "sep.pair"}{ Returns \code{TRUE} if \code{target}
  and \code{current} have the same (\code{u},\code{v}) set. }
}}
\keyword{methods}

