\name{pattern} 
\alias{pattern}
\title{ Chain graph pattern computation }
\description{
  Extracts the pattern of a chain graph.
}
\usage{
pattern(amat) 
}
\arguments{
  \item{amat}{ the adjacency matrix of a chain graph.}
}
\value{
  The adjacency matrix of the chain graph pattern.
}
\references{
  Studeny, M. (1997). A recovery algorithm for chain
  graphs. \emph{Int. J. Approx. Reasoning,} \bold{17}, 265-293.
}
\author{ Zongming Ma and Xiangrui Meng}
\keyword{ graphs }

