\name{show-methods} 
\docType{methods} 
\alias{show-methods}
\alias{show,sep.pair-method}
\title{ Show method for sep.pair class}
\description{
 Method for showing a \code{sep.pair} object.
}
\section{Methods}{
\describe{

\item{object = "sep.pair"}{ Successively prints out the pairs (\code{u},
  \code{v}) and their separator \code{s}. }
}}
\keyword{methods}

